Move loading decorator and adjust console/client
Move with_loading_animation from helper.rich into swanlab/sdk/cmd/utils and update call sites to use utils.with_loading_animation (init.py, login.py). Remove the old helper.rich module and its export. Refactor internal console usage by introducing exported variable `c` and switching print/log helpers to use c.print. Improve client initialization error to include the API key, and tighten prompt_init_mode behavior to raise clear RuntimeError when cloud mode requires interactive login but interactivity/TTY is unavailable. Overall this centralizes the loading decorator in cmd utilities and makes console/client/init logic and error messages more explicit.
