+++ swanlab/sdk/cmd/utils/__init__.py
@@ -5,9 +5,10 @@
 @description: cmd 模块工具函数
 """
 
+from functools import wraps
 from pathlib import Path
 
-from swanlab.sdk.internal.pkg import fs, nrc
+from swanlab.sdk.internal.pkg import console, fs, nrc
 from swanlab.sdk.internal.settings import ROOT_FOLDER, settings
 from swanlab.sdk.typings.cmd import LoginType
 
@@ -32,3 +33,22 @@ def append_gitignore(path: Path):
         # 忽略目录下所有文件
         ignore_content = "*"
         fs.safe_write(gitignore_path, ignore_content)
+
+
+def with_loading_animation(message: str = "Initializing SwanLab...", spinner_name: str = "dots"):
+    """
+    一个使用 rich 包装的加载动画装饰器
+    :param message: 动画旁边显示的文字
+    :param spinner_name: 动画样式（如 'dots', 'bouncingBar', 'point' 等）
+    """
+
+    def decorator(func):
+        @wraps(func)
+        def wrapper(*args, **kwargs):
+            # 使用 rich 的 status 作为上下文管理器包裹函数的执行
+            with console.c.status(message, spinner=spinner_name):
+                return func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
