+++ swanlab/sdk/internal/pkg/client/__init__.py
@@ -49,7 +49,7 @@ class Client:
         # 立即进行首次鉴权并挂载凭证
         login_resp = self._refresh_auth(timeout=timeout, warning=False)
         if not login_resp:
-            raise AuthenticationError("Failed to initialize SwanLab client.")
+            raise AuthenticationError(f"Failed to initialize SwanLab client with API Key '{self._api_key}'")
         # 写入登录响应到上下文，由调用者判断是否需要使用
         scope.set_context("login_resp", login_resp)
